requirements.txt
streamlit>=1.32.0
openpyxl>=3.1.0
pandas>=2.0.0
Pillow>=10.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
